extracted RawValidatorTest component
