+++ packages/playground/src/components/index.ts
@@ -4,3 +4,4 @@ export { ThemeSelector } from './ThemeSelector';
 export { Selector } from './Selector';
 export { ValidatorSelector } from './ValidatorSelector';
 export { SubthemeSelector } from './SubthemeSelector';
+export { RawValidatorTest } from './RawValidatorTest';
